backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
qdrant-client==1.9.1
litellm==1.38.0
python-dotenv==1.0.1
pandas==2.2.2
python-dateutil==2.9.0
PyMuPDF==1.24.4
openai==1.30.1
tenacity==8.3.0
